Fixed instanced data
Previously, when multiple different models were being rendered at once AND the # of entities was increasing, there was a graphical issue that presented as an extra of MODEL A being placed at the location/rotation of an instance of MODEL B.

The cause was the fact that I was calling draw() in Graphics and then immediately calling updateInstanceData. I believe this means that I was uploading different data to the VRAM at the same time that the previous frame was being drawn.