POC component destruction
Individual components can be shot by projectiles, each component has its own health and when it reaches 0 it is physically removed from the ship. A bit of memory is kept in case you want to 'revive' that part, and the mass is recalculated. There is also a graphical effect which shows the part 'degrading' (done in shader)